+++ common/src/main/java/dev/aika/smsn/client/gui/components/SwitchBuilder.java
@@ -1,7 +1,6 @@
 package dev.aika.smsn.client.gui.components;
 
 import dev.aika.smsn.annotation.Components;
-import dev.aika.smsn.annotation.RequiresRestart;
 import me.shedaniel.clothconfig2.api.AbstractConfigListEntry;
 import me.shedaniel.clothconfig2.api.ConfigEntryBuilder;
 import me.shedaniel.clothconfig2.impl.builders.BooleanToggleBuilder;
@@ -20,7 +19,6 @@ public final class SwitchBuilder extends AbstractComponentBuilder<Boolean> {
 
     public SwitchBuilder(ConfigEntryBuilder entryBuilder, Object object, Field field) {
         super(entryBuilder, object, field);
-
         this.annotation = field.getAnnotation(Components.Switch.class);
     }
 
@@ -30,8 +28,7 @@ public final class SwitchBuilder extends AbstractComponentBuilder<Boolean> {
                 .startBooleanToggle(fieldNameKey(), getValue())
                 .setSaveConsumer(this::setValue)
                 .setYesNoTextSupplier(this::yesNoTextSupplier);
-        if (defaultObject != null) builder = builder.setDefaultValue(getDefaultValue());
-        builder.requireRestart(field.getAnnotation(RequiresRestart.class) != null);
+        fieldBuilderInit(builder);
         return builder.build();
     }