+++ common/src/main/java/dev/aika/smsn/mixin/MixinPlatform.java
@@ -1,0 +1,10 @@
+package dev.aika.smsn.mixin;
+
+import dev.architectury.injectables.annotations.ExpectPlatform;
+
+public class MixinPlatform {
+    @ExpectPlatform
+    public static void register() {
+        throw new AssertionError();
+    }
+}