+++ forge/src/main/java/dev/aika/smsn/forge/SMSNPlatformImpl.java
@@ -1,11 +1,9 @@
 package dev.aika.smsn.forge;
 
-import dev.aika.smsn.forge.mixin.MixinInfo;
 import net.minecraftforge.fml.loading.FMLLoader;
 import net.minecraftforge.fml.loading.FMLPaths;
 
 import java.nio.file.Path;
-import java.util.Map;
 
 @SuppressWarnings("unused")
 public class SMSNPlatformImpl {
@@ -13,11 +11,12 @@ public class SMSNPlatformImpl {
         return FMLLoader.getLoadingModList().getModFileById(modId) != null;
     }
 
-    public static Path getConfigDir() {
-        return FMLPaths.CONFIGDIR.get();
+    public static String getModName(String modId) {
+        if (!isModLoaded(modId)) return null;
+        return FMLLoader.getLoadingModList().getModFileById(modId).moduleName();
     }
 
-    public static Map<String, String> getMixinClassNames() {
-        return MixinInfo.getMixinClassNames();
+    public static Path getConfigDir() {
+        return FMLPaths.CONFIGDIR.get();
     }
 }