+++ CHANGELOG.md
@@ -1,9 +1,19 @@
+### Unreleased
+
+**Forge**
+
+- Add disable Mixins features
+
+**Fabric**
+
+- Add disable Mixins features
+
 ### 1.2.10
 
 **Forge**
 
 - Add Mixins for [TenshiLib](https://github.com/Flemmli97/TenshiLib)
-- Add Mixins for [Additional Structures](https://modrinth.com/mod/additional-structures) 
+- Add Mixins for [Additional Structures](https://modrinth.com/mod/additional-structures)
 
 **Fabric**